Splitt ut felles tekster
